TangoTest.cpp: Fix compilation error on Windows
We currently can't compile TangoTest due to

2>c:\\Users\Test\projekte\cppTango\build\install\\include\\tango\pipe.tpp(54): error C2664: "Tango::DevicePipeBlob &Tango::DevicePipeBlob::operator <<(Tango::DevicePipeBlob &)" : cannot convert argument 1 from 'int' to 'short &'
2>        with
2>        [
2>            T=int
2>        ]
2>C:\Users\Test\projekte\TangoTest\TangoTest.cpp(3080): note: Siehe Verweis auf die Instanziierung der gerade kompilierten Funktions-Vorlage "Tango::Pipe &Tango::operator <<<int>(Tango::Pipe &,T &)".
2>        with
2>        [
2>            T=int
2>        ]

It works as is on *nix.

By using the equivalent type short, we can make it work. We do have an
explicit template instantiated for the short type.

Close #18.

+++ TangoTest.cpp
@@ -3076,7 +3076,7 @@ void TangoTest::read_generic_blob_rw(Tango::Pipe &pipe)
         } else {
             std::vector<std::string> de_names{"NO BLOB"};
             pipe.get_blob().set_data_elt_names(de_names);
-            int generic_blob_rw_status = 0;
+            short generic_blob_rw_status = 0;
             pipe << generic_blob_rw_status;
         }
         /*----- PROTECTED REGION END -----*/	//	TangoTest::read_generic_blob_rw
